requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.1
gTTS==2.5.1
idna==3.7
pydantic==2.7.1
pydantic_core==2.18.2
PyPDF2==3.0.0
pypiwin32==223
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
requests==2.31.0
telebot==0.0.5
typing_extensions==4.11.0
urllib3==2.2.1
